+++ packages/core/config/database.php
@@ -15,4 +15,15 @@ return [
     |
     */
     'users_id_type' => 'bigint',
+
+    /*
+    |--------------------------------------------------------------------------
+    | Disable migrations
+    |--------------------------------------------------------------------------
+    |
+    | Prevent Lunar`s default package migrations from running for the core.
+    | Set to 'true' to disable.
+    |
+    */
+    'disable_migrations' => false,
 ];