+++ packages/admin/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `customer` relationship to Cart model
 - Added `setCustomer` method to Cart model
 - Added `setCustomer` and `getCustomer` to StorefrontSessionManager
+- Config to disable database migrations.
 
 ### Fixed