+++ packages/core/CHANGELOG.md
@@ -28,6 +28,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `customer` relationship to Cart model
 - Added `setCustomer` method to Cart model
 - Added `setCustomer` and `getCustomer` to StorefrontSessionManager
+- Added pipelines to PricingManager.
+- Config to disable database migrations.
 
 ### Changed